move assert helper out of dnsdist tests
