+++ regression-tests.dnsdist/eqdnsmessage.py
@@ -1,0 +1,1 @@
+../contrib/assert-equal-DNSMessage/eqdnsmessage.py
